NN WIP; but nuking scheduler now
So I am adding the neuralnetwork support into fiesta. However the scheduler experiment was not very successful back then, so I decided to just completely remove it here. This commit can be used to get the old code if I want to add the schedulers in the future, although that will be a follow up work.