Added LightcurvePlotter and update sys. uncertainty sampling
Added new plotting functionality in fiesta.plot. The LightcurvePlotter
makes plotting lightcurves easy. Also updated sys. uncertainty sampling
so that it a) applies to the non-detections now as well, plus the time
range for specific .yaml entries can be set now, and whether to space
the time nodes log-uniform or not.